requirements.txt
discord.py>=2.3.2
yt-dlp>=2024.3.10
python-dotenv==1.0.1
PyNaCl>=1.5.0